Added provisions for ending a game prematurely
The user may now click the "close window" button provided by the OS for
in an attempt to exit the game in progress.
－If done on an OPTIONAL dialog/action, [see REQUIRED below].
－If done on a REQUIRED dialog/action, ends the action and kicks the
human player out of the game. Remaining CPU opponents finish game and
it stops. A new game can be continued, and the main program does not
fully exit.
－If done on the MAIN MAP WINDOW, accepting will cause the game to full
end. There is no way to deny in this case. The game cannot be restarted
without relaunching the app.

When using the "safe close" and "force close" buttons on the main map
window (not the buttons provided by the OS!), the game cannot be
restarted without relaunching the app. These buttons will also close
dialog windows.