Unfortunate bug in card-turn-in disallowing accept OR skip
-"Turn in cards" patched: previously, the "turn in cards" dialog would
not have run its internal self-check on the response being submitted,
potentially resulting in the game master thinking the user was
attempting to skip by returning an invalid "cardTurnIn" response. The
appropriate logic for the "accept" button was overridden by the button
logic for "skip". Furthermore, the dedicated "Skip" button was left
with no logic assigned, essentially making it a dummy button. The logic
for each button has been correctly wired to said buttons, and the logic
for the "accept" button has been tightened up to indicate when a set of
3 cards is not valid (previously, the window would remain open, but no
error message would appear).

-(miscellaneous layout/naming/cleaning changes were also put into
place).