Code layout/naming tweaks, small logic tweaks
－Fix: Main map UI refreshing: In certain circumstances, when only CPU
bots are left playing, the UI refresh method would cause a
concurrentModificationException to be thrown and make the UI lag until
the bots completed. This update attempts to circumvent that issue.
－Tweak: Main map UI refreshing: the UI would always try to refresh
slowly, even when the bots are speeding through their changes,
potentially causing lag. This update attempts to detect when a bot is
playing to force the UI thread to immediately update (or immediately
queue an update, as required). (More work required to improve quality
of fix).

－Naming convention: select variables and functions have been re-named
in an attempt to be more indicative of what they do or what they are
used for.

－Comments: select comments have been beefed up, or updated from old
copy/paste versions. Essentially, some comments are more relevant now.

－Code layout: select portions of code have been re-flowed to allow as
much initial instantiation to occur at the beginning of a method as
possible. This is an attempt to improve readability & tweakability in
the future. (Any major issues that might occur after this update likely
stem from this change).

－Very tiny, tiny, minor reduction in redundancy in select spots.

－Very tiny front-facing UI tweak to make window auto-sizing more
accurate with cardTurnIn dialog.