More renaming to fit standards, preparation for slight logic tweaks when exiting &etc
Internal logic paths should not be massively affected with this change.
However…

－Restored the ability for a user to choose whether to exit a game or
not using the default OS decorations on the various response dialogs
(you are no longer required to use the "end game" button on the main
window). The logic for this is somewhat captive, so if a dialog refuses
to go away, that’s probably the reason.

－Various variable names have been tweaked, and use of select variables
has been cleaned up by adhering to standards re: accessing static class
members. This has been done to facilitate clarity & future upgrades.