selectively enabled game-end confirm dialogs
Did my best to add a test setup to restore confirmation dialogs when a
user attempts to end a game session without closing the application. The
confirmation upon trying to fully exit the app (which was not actually a
confirmation, so much as a notification) has not been re-enabled.
Feature parity with FXUIGM version 11 should be restored, with the added
benefit that the application should no longer block OS commands
(politely asking the app to hide itself or completely close failed when
using the OS X dock, specifically).

--The confirmation/indication window has been tweaked to improve
awareness of its existence, with more changes to come.

--Code spacing/tabbing may be a mess at the moment...or it may be
flawless. Depends on what editor/IDE is in use.