Refinement of logic, quirks fixed.
-Altered points at which saves/checkpoints are created. This helps
ensure that multiple loads of the same game save do not result in wide
variances in how the game will look upon restore/load. Saving is now
performed only after one full turn is completed by a human (FXUIPlayer).

-When creating dialogs, logic would not always wait long enough for a
window to be displayed, causing a "Do you want to exit?" verification
dialog to be displayed. Didn’t break the game, but was still a quirk.
The delay has been increased to allow for the JFX "window is showing"
flag to be set, appropriately blocking the game logic at the proper
place.

-Previously, trying to load a game resulted in players being set up as
if it was a new game, with the players then being overwritten from the
loaded checkpoint. This redundancy has been reduced.

-Loading is now done on the main game thread, rather than being forced
to the FX thread. This forces sequential actions and provides enough of
a delay to avoid concurrency issues in certain stress-test cases.

-The Fortification response (FortifyResponse) now has a slightly more
robust method to verify if a given f-rsp is valid. (Previously could
error out and cause a stall if null values were in place in the
response).

-Crossbar: returned to a simple version of the method used to check
whether a human is ending the game. Now much more versatile, with no
expected negative impact to prior use.

-Created method to simplify calls to Thread.sleep() within the two main
FXUI classes. (Must be consolidated in the next push).

-Select comments improved/updated.

-Select methods did not indicate an error state whatsoever, despite the
error checking logic being available. A basic bandage has been applied
in these scenarios.

-Simplified one line of logic in multiple FXUIPlayer methods. [Improved
readability.]

-Tweaked style in which the application prepared to exit － and
subsequently exited. The verification dialog should be fully enabled
once again.

-Attempting to exit the game in a manner which triggers the
"Application is exiting…" dialog no longer results in the "Do you want
to end the game?" dialog being displayed thereafter. (The current patch
is merely a workaround, instead of a fix).

Known issues:
-Occasionally produces an error upon close, indicating resources are
being accessed even with the JVM being closed, or a Thread has its
parent terminated without the Thread in question being terminated
itself.