Bugs(?) and App-exit tweaks.
－The window to turn cards had the logic for a "skip" button, but the
button was not placed in the layout. This has been fixed. (Button is
disabled when the action is required, but is still visible in the
dialog at all times).

－Trying to exit the game…is probably still broken (in that there will
probably be some silent errors). But now it’s exacerbated by a
completely needless dialog! Gaze in wonder at it! Hopefully you’ll
enjoy. =)

－More attempts to help short-circuit the logic when a game-close
condition is raised. This must be refined in the future.

--Reformatted the loadFromSave() method, bypassing the old conditional
logic that had previously been in place.

－Preparation for more detailed info when the act of loading a game
fails. (Text is stored, but there is no current way to view it).
Additional work will be done in the future, and the same will be done
for the act of saving a game.

－The Crossbar now stores a copy of the names of all human Players (all
FXUIPlayers). Again, no point to this now, but there for further use in
the future (specifically for potential multiplayer, and differentiating
between say ending a game altogether vs one player forfeiting).

－Moved the "sleep" method into RiskUtils to simplify the FXUI classes a
tiny bit. (Previously, each class has its own copy of the method).

－Rare bug where the "turn in a card?" dialog will be re-displayed,
potentially due to misinterpretation as an invalid response rather than
a "skip" triggered by the use of the "Accept/OK" button. I haven’t
fixed the bug or anything, but I’m letting you know it’s there, and I’m
letting you know I have no known cause for this at the moment.