Avoids crash on exit, misc. tweaks.
Ameliorated: Frequently, trying to exit the game stalled due to
FXUIPlayer refusing to stop waiting for an event that would never
happen. Well, the singular thread stalled. Everything else closed fine
but that thread, which was YANKED INTO THE DEPTHS BELOW I mean forced
to terminate in a displeasing fashion. That chance of that happening
should be reduced…though the thread may occasionally still slow down
the entire program exit. (Should the game lock up while trying to exit,
it’s due to this fix…well, one part of it).

Tweaked: miscellaneous naming and etc.