Support custom saves/loads.
Basic support for custom save file locations. The game will use this
new location, once set, as the auto-save location until restarted.
Otherwise, the auto-save location will always be the default. Setting
does not persist with save files.

Basic support for loading game saves from a different location. The
game will treat this as the "default" location until restarted. Does
not affect the autosave location. Setting does not persist with save
files.

You are now prompted with the ability to load from a previous file or
start a new game. The game will pre-check for a default checkpoint from
which you can load and offer that, if available; else, to load, you
will have to select a game file. New games are always available.

Tooltips are used in various places to hold more status information. Be
sure to check them!

Other tiny bits of the layout have been tweaked