Minor usability tweak.
Enabled: ability to click on a player's name in the main window to
highlight owned countries during game play. (Previously, this was
limited to the current player via a dedicated button).

Tweaked: Updated text on the button used to load a game.

Tweaked: Status info on the main window tweaked to add character or be
more informative.

Tweaked: Logic for enabling/disabling select buttons on the main window
makes use of an ENUM setup, to make code more explicit in what it's
doing. Also, logic to set up the button cache has been updated to make
better use of the ENUM variables.

Fixed: Logic no longer includes error checking during initial launch
setup, as most of the error checking has been relegated to later in the
execution process. (Errors are still indicated through the various
status text, or in their respective dialogs.)