Fix build script, add clean script
Fixed: An obscure error regarding String instance variables declared as
final occurs with JDK8 or JDK9, resulting in the secondary dialogs
failing to open (essentially locking the user out of proceeding with
their turn). To get around this, added the "-g" flag to the call to
javac (the compilation stage) which enables all debugging information
within the final code, which works around this. The cause of this issue
is unknown, but does not occur with JDK7.

Tweaked: text sent to the standard output directly from the script is
indented, to provide a visible difference from messages printed by the
called processes themselves.

Tweaked: references to files are more specific, where required, to
avoid copying useless/unnecessary files as resource files. Also added a
few more comments and tweaked some stdout/echo text.

Added: cleaning script, as a companion to the makejarjfx script. (Also
removes the generated Jar file!); essentially attempts to return the
folder to the original state, minus any extra files produced during the
run of the application itself.