Resizing perf tweaks & known bugs
[warning: take heed of the "diagnosticModeEnabled" boolean within the
class, as it is set to 'true' in this commit and will result in quite a
bit of debugging info flooding your command line/terminal.]

Reduced the delay between steps in the window resizing process, as the
earlier delay target resulted in a lack of responsiveness. The delay
may be shortened or otherwise tweaked again in the future.

Fixed an issue where the code meant to keep the window from expanding
beyond visual bounds would not be run on the proper thread, potentially
resulting in distortion of the content and/or the window itself.
(Internally, this could be detected as an unchecked exception being
thrown to alert the developer and prevent a fatal error, which is by
design.) Other aspects of the resizing code were not affected by this
fix.

Added a little code to further reduce the chance of operating on a null
object, or attempting to operate while the associated object had been
marked as inactive. More must be done to support this new "inactive"
flag.

Known bug: switching to fullscreen then switching back causes improper
content scaling, resulting in an irregular border appearing. The
current workaround is to resize the window every time you exit
fullscreen mode.

Expected bug: going into fullscreen results in the aspect ratio not
being maintained, with the content being stretched or compressed to fit
the screen perfectly (sometimes with irregular cropping of the
content). No workaround exists. (A window that has been maximized is
your best bet in this case).