Fix bug which caused hang on exit
If the "new game/load game" screen was showing & the user attempted to
quit, in some circumstances the game would pause during its exit, with
a blank "good bye" window (though the executable was technically still
responsive). A second attempt to close would result in the game
completing its exit. This has been resolved through some restructuring.
This could be specifically seen with a key combo like CMD + Q in OS X,
where the keypress would cause the game to attempt to launch, only to
have the selector window bug summarily prevent the app from fully
exiting without pressing CMD + Q again.

Other miscellaneous changes were implemented, and the shell for the
persistent settings was removed. (Game will continue to default as it
already does, and a few more options may be added to the
SavePoint/checkpoint feature as desired.)