rough implementation of auxiliary messages
rough implementation of auxiliary messages: testing the ability to
display tertiary status messages on the screen, as an additional
status/error message space (to display long messages). If the message
is over a certain length of characters, it can be displayed in a
separate dialog by clicking on the text that is visible.