Enable display of log text during game play
[*] Added the ability to click a button to display text re: current
game activity.

－This allows you to read every move taken by the opponent bots, as well
as to read the moves you took. The original visual LogPlayer has NOT
been made redundant by this, as this "log viewer" is for the active
game ONLY. Reviewing the logs of past games remains the purpose of the
original visual LogPlayer.

－Note: With this update comes a revision to the SavePoint class, due to
a switch in internal data structures. (A switch from explicit use of
ArrayLists in select spots to an implicit use, with no explicit casts).
Although this is not expected to cause issues (hence the internal
SavePoint serial version remaining the same, allowing unpacking of
earlier saves), please back up your previous game saves in case of
corruption. This save format may be artificially archived in the near
future, depending on the need to save other user options.

[*] Added: The ability to watch bots play one another. (You cannot yet
control how many bots, nor can you control what type.) Watching bots
play from the get-go will result in a 1-second delay between player
moves. This will be revised in the future.

[*] Fixed: inability to end game when bots were playing: in earlier
versions, when the bots were making decisions between the human’s
moves, it is possible that trying to end the game without exiting the
application could result in the human being ignored until the next time
a player dialog appeared. This has been fixed for both that old
scenario and for the new "bots only play" scenario.

[****] This feature set will be frozen & debugged for a proper BETA or
FINAL release (currently in ALPHA, been that way for forever now).
Uploads may be few and far between.