The bug-fix special.
-Many comments updated. Because…necessary.

-Added: ability to see the status of the saving process

-Refined: select text elements have been resized.

-Added: activity indicators have been added to the main screen

-Added: indicators of the current player have been added to the main
screen

-Added: audio cues and jingles. It’s rough, but it’s there.

-Changed: SavePoints now contain the functions to save themselves.
Hahaha. But seriously, the method was moved from FXUIGameMaster to the
SavePoint class to save a little space.

-Added: an icon? But it’s not used as an application icon. It’s really
more of a logo that’s the equivalent of a 2-minute Photoshop job.

-Tweaked: the About and More Info dialog code is now in its own class
file (previously resided alongside FXUIGameMaster)

-Tweaked: Options and About buttons are now grouped together in one
button on the main UI.

-Tweaked (well, really, just added): Audio options exist (mute, volume)

-Tweaked: Window options now function more cohesively and reliably.

-Fixed: long-standing bug resulted in some platforms having window
decoration that was ignored during window resizing operations, meaning
the content wasn’t scaled properly and potentially was truncated. This
has been fixed.

-Tweaked: window resizing operations in general should be more reliable
when transitioning into and out of fullscreen mode on some platforms.

-Tweaked: some dialog windows (not related to the FXUIPlayer, yet) have
different layout properties.

-Tweaked: dialog auto-close timeouts have been increased slightly. The
About screen shown at launch also allows you to delay the auto-close by
hovering your mouse.

-Hopeful optimization: UI refreshes related to countries should now be
slightly less taxing due to specific instructions on what must be
updated and how (whether it was an owner change or an army count change)

-tweaked: some odd number of variables were renamed for clarity.

-added: ability to control the delay between moves done by bots

-oh, added: ability for bots to play by themselves.

-tweaked: some tooltips and instructions have been edited for clarity,
or to remove /some/ of the cold, uncaring robotic feeling that
previously existed.

-added: boot and shutdown splash screen animations.

-tweaked: beginning to hide elements rather than disabling them in some
scenarios.

-tweaked: optimized some threaded activities to reduce creation of new
threads.

-tweaked: in the event of a hang somewhere in the program while trying
to exit, the program may not-so-gracefully force itself to close
regardless.

-added: ability to view the contents being written to the text log.
This is different from LogPlayer (which gives you a visual and textual
representation replay of a prior game; this gives you written text of
the game in progress)

-tweaked: UI updates are more consistent and should occur with each
action taken by bots (previously, UI refreshes were only guaranteed
when the human player was playing, and bot UI refreshes could be
delayed)

-Fix: trying to end the game during bot play could have failed
previously. This has been resolved.