Bug prevention & bug fixes.
-tweaked: win states no longer animate the player's countries (which is
the entire map), as it was superfluous and slowed down animation of the
"you have won!" ribbon. Other aspects of a win state have been tweaked
re: status text/info.

-fix part 1 of ?: in some circumstances, the method used to highlight
the current player could cause ConcurrentModification exceptions. The
method has been reflowed to reduce the change that the method will cause
the issue for itself. Other aspects of the fix will have to be
implemented elsewhere. The patch implemented here will have the code to
retry an attempt that causes such an exception, in the hopes that the
failed example of a race condition will have cleared up.

-fixed: loading of the audio manager now has a reduced chance of
delaying application boot time/making the initial splash screen
animation choppy/making the About window refuse to show upon app boot.
The audio files are now loaded in a separate thread of sorts.

-tweaked: the boot sound now plays upon load of the audio manager class,
no separate call required.

-tweaked: setting the volume now gives you a numeric representation of
the current percentage (rather than merely estimating based on the
position of the slider)