Continued fixes and improvements
-Added (finally): ability to select how many cars you use to defend
yourself from an attack. Previously, you weren’t notified, and it
automatically just replied with the greatest risk/greatest reward
strategy (rolling the maximum dice available)

-Fixed: selecting the number of dice to perform an attack should no
longer be broken. Not sure how that ended up being a problem, but
simple mistakes…

-Under investigation (maybe patched?): sometimes, putting an extended
message in the lower left of the screen fails. Now a larger chunk of
the code to do so is guaranteed to run on the JavaFX thread, in a
half-hearted attempt to reduce the potential catastrophe that could
come from trying to set a UI element on a plain thread.

-Slightly ameliorated: animations should flicker less often
(previously, it was possible for two threads to try to run the
animation code simultaneously).

-Possibly ameliorated: Some HashMaps have been wrapped up as
synchronized Map objects to reduce the chance of
concurrentModificationExceptions. The downside is that atomic
operations could potentially result in operations being out-of-order,
meaning screen refreshes (or similar) could be delayed.

-Tweaked: some animations may be faster and/or smoother. *shrug*

-Tweaked: layout of screen when attacking a country has been fixed so
that elements are more centered.

-Tweaked: Bash build script has been tweaked to ask the user if they
want to delete extraneous build files after running (or attempting to
run) the Jar created.