Misc bug fixes & workarounds
This release was simultaneously developed on Eclipse and Netbeans,
hence some of the fixes/workarounds.

Workaround: in some cases, loading audio files will fail, even though
the files are in the jar file, due to a different interpretation of the
file structure & build process between Eclipse, Netbeans, and the
custom Bash script.

Workaround: in some cases, the LogPlayer wouldn’t find the
TextNodes.txt file, and thus couldn’t finish loading its display. The
workaround implemented some months back for FXUIGameMaster (which looks
to the TextNodes class for the list of countries) has been applied
here. Of note is that if the countries are of a generic color (black or
white) upon loading of either FXUIGM or LogPlayer, then the app is
using the text file. If it is purple/blue, it is using the TextNodes
class.

－Tweaked: abusing Netbeans’ automatic fixes and improvements to make
use of Java 8’s functional operations. (Used over traditional for-loops
or iterators where suggested)

Misc fixes and unnecessary bits taken out at the suggestion of Netbeans.

－Changed: selecting "bots only play" for simulation will load 2x easy,
2x normal, and 2x hard players. Selecting "human + bots" is like
before, with 1x of easy, normal, and hard bots each. There is currently
still no way to select a Seth bot or any other bot player, or do a
mix/match kind of thing.