Swap "hover" for "long press" at times
In the last release, some buttons were activated by a long hover, and
repeatedly took action so long as the mouse was hovering over the
button. These buttons now require a press (single click, down) to
activate in a similar manner (you may press then hover to have the
button continue to activate, OR you may press & hold to have the button
continue to activate). To deactivate:
－mouse users: release the mouse key and move the mouse cursor away from
the button.
－touchscreen users: tap away from the button

Future versions of this software may make this an option (click to
activate, or long hover) available directly to the user.