minor fixes
Works around two minor issues:
－Switching to fullscreen with dialogs visible tended to hide the
dialogs, with no way to bring them forward again. The app now attempts
to keep the main window as far back as possible during the transition,
leaving the dialogs in the front.

－Switching to another application while the game was booting (before
the About screen disappeared, specifically) would result in the game
stealing focus once the About screen disappeared. The game now attempts
to send the main window to the back of all other open windows right as
the About dialog closes IF the window is not in focus. This should
reduce annoyance.