begin implementation of autobrite function
As the vague title says, added the first stages of an automatic
brightness function. It functions based on the time of day (or can be
manually set). In tandem with this is a yellow filter to help reduce
eyestrain or make the screen more pleasant to look at in dark
environments (this is an on/off option with no range).

These options are tucked under Window Options or Window Size Options,
depending on how it is currently labeled.

These options currently affect ONLY the main window, and do not affect
secondary dialog windows. This will be added in the future.